# product-api/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    GenericError:
        description: GenericError is a generic error message returned by a server
        properties:
            message:
                type: string
                x-go-name: Message
        type: object
        x-go-package: github.com/nicholasjackson/building-microservices-youtube/product-api/handlers
    Product:
        description: Product defines the structure for an API product
        properties:
            description:
                description: the description for this poduct
                maxLength: 10000
                type: string
                x-go-name: Description
            id:
                description: the id for the product
                format: int64
                minimum: 1
                type: integer
                x-go-name: ID
            name:
                description: the name for this poduct
                maxLength: 255
                type: string
                x-go-name: Name
            price:
                description: the price for the product
                format: float
                minimum: 0.01
                type: number
                x-go-name: Price
            sku:
                description: the SKU for the product
                pattern: "[a-z]+-[a-z]+-[a-z]+"
                type: string
                x-go-name: SKU
        required:
            - name
            - price
            - sku
        type: object
        x-go-package: github.com/nicholasjackson/building-microservices-youtube/product-api/data
    ValidationError:
        description: ValidationError is a collection of validation error messages
        properties:
            messages:
                items:
                    type: string
                type: array
                x-go-name: Messages
        type: object
        x-go-package: github.com/nicholasjackson/building-microservices-youtube/product-api/handlers
info:
    description: Documentation for Product API
    title: of Product API
    version: 1.0.0
paths:
    /products:
        get:
            description: Return a list of products from the database
            operationId: listProducts
            responses:
                "200":
                    $ref: "#/responses/productsResponse"
            tags:
                - products
        post:
            description: Create a new product
            operationId: createProduct
            parameters:
                - description: |-
                      Product data structure to Update or Create.
                      Note: the id field is ignored by update and create operations
                  in: body
                  name: Body
                  required: true
                  schema:
                      $ref: "#/definitions/Product"
            responses:
                "200":
                    $ref: "#/responses/productResponse"
                "422":
                    $ref: "#/responses/errorValidation"
                "501":
                    $ref: "#/responses/errorResponse"
            tags:
                - products
        put:
            description: Update a products details
            operationId: updateProduct
            parameters:
                - description: |-
                      Product data structure to Update or Create.
                      Note: the id field is ignored by update and create operations
                  in: body
                  name: Body
                  required: true
                  schema:
                      $ref: "#/definitions/Product"
                - description: The id of the product for which the operation relates
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "201":
                    $ref: "#/responses/noContentResponse"
                "404":
                    $ref: "#/responses/errorResponse"
                "422":
                    $ref: "#/responses/errorValidation"
            tags:
                - products
    /products/{id}:
        delete:
            description: Update a products details
            operationId: deleteProduct
            responses:
                "201":
                    $ref: "#/responses/noContentResponse"
                "404":
                    $ref: "#/responses/errorResponse"
                "501":
                    $ref: "#/responses/errorResponse"
            tags:
                - products
        get:
            description: Return a list of products from the database
            operationId: listSingle
            responses:
                "200":
                    $ref: "#/responses/productResponse"
                "404":
                    $ref: "#/responses/errorResponse"
            tags:
                - products
produces:
    - application/json
responses:
    errorResponse:
        description: Generic error message returned as a string
        schema:
            $ref: "#/definitions/GenericError"
    errorValidation:
        description: Validation errors defined as an array of strings
        schema:
            $ref: "#/definitions/ValidationError"
    noContentResponse:
        description: No content is returned by this API endpoint
    productResponse:
        description: Data structure representing a single product
        schema:
            $ref: "#/definitions/Product"
    productsResponse:
        description: A list of products
        schema:
            items:
                $ref: "#/definitions/Product"
            type: array
schemes:
    - http
swagger: "2.0"